+++ playtime/src/web-routes.js
@@ -1,5 +1,6 @@
 import * as accounts from './controllers/accounts-controller.js';
 import * as dashboard from './controllers/dashboard-controller.js';
+import * as playlist from './controllers/playlist-controller.js';
 
 const webRoutes = [
   { method: 'GET', path: '/', options: { auth: false }, handler: accounts.index },
@@ -9,6 +10,8 @@ const webRoutes = [
   { method: 'POST', path: '/authenticate', options: { auth: false }, handler: accounts.login },
   { method: 'GET', path: '/logout', handler: accounts.logout },
   { method: 'GET', path: '/dashboard', handler: dashboard.index },
+  { method: 'POST', path: '/dashboard/addplaylist', handler: dashboard.addPlaylist },
+  { method: 'GET', path: '/playlist/{id}', handler: playlist.index },
 ];
 
 export default webRoutes;